Bug fixed for using with SOEM
The bug encountered when using SOEM has been fixed.
In the generated .ino file, PDO mapping entries (eg. 1600, 1A00 ... ) should be of type OBJCODE_ARR. However, it used to be OBJCODE_REC. This caused the problem that when the master issues a command of SDO_UPLOAD using complete access, the slave could not calculate the correct object length. Only the length of the first entry is calculated, but the exact length should be calculated based on sub-index 0.